requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.20.0
gspread>=6.0.0
google-auth>=2.28.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
